some changes in app/todoList
